Differ between real and temporary exit values when computing result
The Icinga API only accepts status 0 or 1 for host objects and 0,1,2,3
for service objects. But when a warning or critical threshold is
breached for a host object the plugin output should say so, even
though it ultimately sends either exit status 0 or 1 to the API.
The distinction between temp and "real" exit value is only used to
compute better default plugin output. The "exit_value" used in the
handlebars templates will be the "real" one, so 0-3 for services and 0
or 1 for hosts.